# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FlowCI API
  description: Docker image build and deployment API
  version: 0.1.0
  contact:
    name: FlowCI Team
servers:
  - url: http://localhost:3847/api/v1
    description: Development server
paths:
...
  /builds:
    get:
      summary: List all builds
      operationId: listBuilds
      tags:
        - Builds
      responses:
        '200':
          description: List of builds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildListResponse'
    post:
      summary: Create a new build
      operationId: createBuild
      tags:
        - Builds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuildRequest'
      responses:
        '201':
          description: Build created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /builds/{id}:
    get:
      summary: Get build by ID
      operationId: getBuild
      tags:
        - Builds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Build details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildResponse'
        '404':
          description: Build not found
  /builds/{id}/logs:
    get:
      summary: Get build logs
      operationId: getBuildLogs
      tags:
        - Builds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Build logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildLogsResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
...
    BuildResponse:
      type: object
      properties:
        id:
          type: string
        image_id:
          type: string
        tags:
          type: array
          items:
            type: string
        size:
          type: integer
        duration_ms:
          type: integer
        status:
          type: string
        logs:
          type: array
          items:
            type: string
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
    BuildListResponse:
      type: object
      properties:
        builds:
          type: array
          items:
            $ref: '#/components/schemas/BuildResponse'
        total:
          type: integer
    BuildLogsResponse:
      type: object
      properties:
        build_id:
          type: string
        logs:
          type: array
          items:
            type: string
...